Fix docs build
