+++ flash/data/process.py
@@ -13,8 +13,7 @@
 # limitations under the License.
 import os
 from abc import ABC, abstractclassmethod, abstractmethod
-from dataclasses import dataclass
-from typing import Any, Callable, Dict, List, Mapping, Optional, Sequence, Type, TYPE_CHECKING, TypeVar, Union
+from typing import Any, Callable, Dict, List, Mapping, Optional, Sequence, TYPE_CHECKING, TypeVar, Union
 
 import torch
 from pytorch_lightning.trainer.states import RunningStage
@@ -25,111 +24,12 @@ from torch.utils.data._utils.collate import default_collate
 
 from flash.data.batch import default_uncollate
 from flash.data.callback import FlashCallback
+from flash.data.data_source import DataSource
+from flash.data.properties import Properties
 from flash.data.utils import _PREPROCESS_FUNCS, _STAGES_PREFIX, convert_to_modules
 
 if TYPE_CHECKING:
     from flash.data.data_pipeline import DataPipelineState
-    from flash.data.data_source import DataSource
-
-
-@dataclass(unsafe_hash=True, frozen=True)
-class ProcessState:
-    """
-    Base class for all process states
-    """
-    pass
-
-
-STATE_TYPE = TypeVar('STATE_TYPE', bound=ProcessState)
-
-
-class Properties:
-
-    def __init__(self):
-        super().__init__()
-
-        self._running_stage: Optional[RunningStage] = None
-        self._current_fn: Optional[str] = None
-        self._data_pipeline_state: Optional['DataPipelineState'] = None
-        self._state: Dict[Type[ProcessState], ProcessState] = {}
-
-    def get_state(self, state_type: Type[STATE_TYPE]) -> Optional[STATE_TYPE]:
-        if state_type in self._state:
-            return self._state[state_type]
-        if self._data_pipeline_state is not None:
-            return self._data_pipeline_state.get_state(state_type)
-        else:
-            return None
-
-    def set_state(self, state: ProcessState):
-        self._state[type(state)] = state
-        if self._data_pipeline_state is not None:
-            self._data_pipeline_state.set_state(state)
-
-    def attach_data_pipeline_state(self, data_pipeline_state: 'DataPipelineState'):
-        self._data_pipeline_state = data_pipeline_state
-        for state in self._state.values():
-            self._data_pipeline_state.set_state(state)
-
-    @property
-    def current_fn(self) -> Optional[str]:
-        return self._current_fn
-
-    @current_fn.setter
-    def current_fn(self, current_fn: str):
-        self._current_fn = current_fn
-
-    @property
-    def running_stage(self) -> Optional[RunningStage]:
-        return self._running_stage
-
-    @running_stage.setter
-    def running_stage(self, running_stage: RunningStage):
-        self._running_stage = running_stage
-
-    @property
-    def training(self) -> bool:
-        return self._running_stage == RunningStage.TRAINING
-
-    @training.setter
-    def training(self, val: bool) -> None:
-        if val:
-            self._running_stage = RunningStage.TRAINING
-        elif self.training:
-            self._running_stage = None
-
-    @property
-    def testing(self) -> bool:
-        return self._running_stage == RunningStage.TESTING
-
-    @testing.setter
-    def testing(self, val: bool) -> None:
-        if val:
-            self._running_stage = RunningStage.TESTING
-        elif self.testing:
-            self._running_stage = None
-
-    @property
-    def predicting(self) -> bool:
-        return self._running_stage == RunningStage.PREDICTING
-
-    @predicting.setter
-    def predicting(self, val: bool) -> None:
-        if val:
-            self._running_stage = RunningStage.PREDICTING
-        elif self.predicting:
-            self._running_stage = None
-
-    @property
-    def validating(self) -> bool:
-        return self._running_stage == RunningStage.VALIDATING
-
-    @validating.setter
-    def validating(self, val: bool) -> None:
-        if val:
-            self._running_stage = RunningStage.VALIDATING
-        elif self.validating:
-            self._running_stage = None
 
 
 class BasePreprocess(ABC):
